Fix triggered requests without persistence
This involved a major refactor of how the template engine interfaces with the HTTP engine. An added layer of abstraction allows the CLI and TUI to control this behavior independently.
