+++ Cargo.toml
@@ -27,6 +27,7 @@ rust-version = "1.85.0"
 # Dependencies used in multiple crates get de-duplicated here
 [workspace.dependencies]
 anyhow = "1.0.0"
+async-trait = "0.1.81"
 bytes = {version = "1.6.1", default-features = false}
 chrono = {version = "0.4.31", default-features = false}
 crossterm = {version = "0.28.0", default-features = false, features = ["events"]}
