+++ crates/core/src/http/models.rs
@@ -345,11 +345,6 @@ pub struct RequestTicket {
     pub(super) client: Client,
     /// Our brave little astronaut, ready to be launched...
     pub(super) request: Request,
-    /// Should the request/response be saved to the DB? This is based on:
-    /// - Global config `persist` field
-    /// - Recipe `persist` field
-    /// - CLI `--persist` flag
-    pub(super) persist: bool,
 }
 
 impl RequestTicket {
@@ -360,8 +355,9 @@ impl RequestTicket {
 
 /// A complete request+response pairing. This is generated by
 /// [RequestTicket::send] when a response is received successfully for a sent
-/// request.
-#[derive(Debug)]
+/// request. This is cheaply cloneable because the request and response are
+/// both wrapped in `Arc`.
+#[derive(Clone, Debug)]
 #[cfg_attr(any(test, feature = "test"), derive(PartialEq))]
 pub struct Exchange {
     /// ID to uniquely refer to this exchange
