+++ crates/tui/Cargo.toml
@@ -12,6 +12,7 @@ version = {workspace = true}
 
 [dependencies]
 anyhow = {workspace = true}
+async-trait = {workspace = true}
 bytes = {workspace = true}
 chrono = {workspace = true}
 cli-clipboard = "0.4.0"
@@ -44,7 +45,9 @@ uuid = {workspace = true}
 [dev-dependencies]
 pretty_assertions = {workspace = true}
 rstest = {workspace = true}
+serde_json = {workspace = true}
 slumber_core = {workspace = true, features = ["test"]}
+wiremock = {workspace = true}
 
 [package.metadata.release]
 tag = false
